Refactor: utils.py로 통합 및 새로운 함수 추가
- document_loaders.py를 utils.py로 통합하여 코드 구조 개선
- 용도에 맞는 프롬프트를 반환하는 get_prompt 함수 추가
- URL을 입력하면 해당 사이트의 내용을 불러오는 get_web_loader 함수 추가
